// src/shared/services/api/models/assessmentOverviewDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */
import type { AssessmentCourseDto } from './assessmentCourseDto';

export interface AssessmentOverviewDto {
  id?: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  total_mark?: number;
  passing_mark?: number;
  duration_minute?: number;
  course?: AssessmentCourseDto;
}