// src/shared/services/api/models/questionDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */
import type { RefTermDto } from './refTermDto';

export interface QuestionDto {
  id?: string;
  assessment_id?: string;
  meta_topic_id?: string;
  question_type?: RefTermDto;
  /** @nullable */
  question_text?: string | null;
  /** @nullable */
  option_list?: string[] | null;
  /** @nullable */
  answer?: string | null;
  weight?: number;
}