// src/shared/services/api/models/updateQuestionRequestDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */

export interface UpdateQuestionRequestDto {
  /** @nullable */
  question_text?: string | null;
  /** @nullable */
  option_list?: string[] | null;
  /** @nullable */
  answer?: string | null;
  /** @nullable */
  meta_topic_id?: string | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  question_type_id?: string | null;
}